Moved swapchain to its own package
Further refactors are needed for better encapsulation and separation of concerns. Currently, the rendering context has to take care of re-creating the swapchain. This is problematic as teardown for the swapchain components needs to happen in specific order and should thus be encapsulated inside the swapchain wrapper itself, not the rendering context.

That is, further down the line, the SwapChain class should be all that is needed for creating and re-creating swapchains.