Cleaned up the application and device init
This was done in preparations for moving to dynamically generated meshes. The Application.java was starting to get so messy, working with it was practically impossible. Now, all functionality is nicely wrapped in helper methods etc.

Also, poked the DeviceContext.java a bit as it was one of the first things written using MemoryStack and did not properly utilize the stack. (or rather, was quite inefficient at that) Now the device context is initialized via standard constructor, like all other instance and context classes.

The ApplicationContext now accepts an array of vertices and indices to allow hard-coding the vertex input at higher level. This should be enough for early iterations of the path generator, although vertex formats still need tweaking to fully support 3d coordinates.