Fixed cave not being correctly centered
The issue was with how the sum of node positions for the average was calculated. The sum was calculated using the offsets x, y and z, which happen to be more or less uniformly distributed, resulting in averages of approximately zero with larger caves.

Fixed by changing the sum to use the absolute location of the generated node and the cave is now properly centered!