[wip] Refactored render pass out of pipeline
The render pass being part of pipeline was conceptually wrong to begin with. In reality, it's more the other way around; the pipeline belongs to a render pass.

But why is this necessary now? The pipeline having the ownership of the render pass prevents re-using it for other pipelines. Then again, one pipeline can only have a single VK_PRIMITIVE_TOPOLOGY_XXX which cannot be changed without recreating the pipeline (which we REALLY do not want to make, trust me). Now, with the render pass moved to upper level, we can create multiple graphics pipelines and ultimately overlay multiple visualizations with different primitive types. e.g. LINE_STRIP for path and POINT_LIST for density samples.

TL;DR: This allows nicer toggling between visualizations in the near future.