Add explicit point size
Vulkan specification requires that when rendering points, the point size must be stated via PointSize output variable in the vertex shader. It defaults to 1.0 on practically all drivers, but as not specifying point size is undefined behavior by the spec, its better to be safe than sorry.

This ceases the log spam from the corresponding validation layer error.