Depth/framebuffers moved to the RenderPass
They were not used anywhere else, so it seemed logical step to refactor their initialization to the render pass. Also, this moved a lot of clutter out of the RenderingContext. The latter was the actual goal behind this, as I didn't want to start adding more stuff like vertex formats etc. before the initialization order was clearer.