Random Walk now has directional bias
Previously, the new direction vector for each step had more-or-less uniform distribution. That is, stepping with 180 degrees turns was as likely as continuing straight ahead. This resulted in some very ugly-looking knots in the generated path.

To overcome the issue, the walk now remembers the direction it last took step towards. The new direction values are summed to this old direction which is then normalized to step length. This gives the algorithm a noticeable bias and prevents it from taking too steep U-turns. This likely still requires some tweaking but initial results are promising.