Cave now has flat floor
It was kind of starting to hurt my eyes that the cave was perfectly round. Changed the density function to use zero-clamped dot product between up-vector and direction from position to closest point on path as weight value for multiplying the density. This sounds complex but is actually quite simple trick for adjusting density function based on whether or not the position is on floor or ceiling.

This *might* cause issues later on if overall cave shape changes to be more complex as this will attempt to flatten all and any upwards facing surfaces. The weight value can quite easily be squared etc. to mitigate the negatives though, but I'm not too worried about that for now.

This was mostly a proof-of-concept as I'd like to overlay different noise functions later on and having everything equally noisy might not look too good.