Implemented a profiler utility
Added SLF4J as dependency and wrapped it to a nice Profiler utility for cleaner logging of profiling information. This allows us to print nice formatted summary of what is going on during the initialization, and is more future-proof as we got rid of all ugly SOUTs and SERRs. Also, as time handling is baked into the profiler utility, we get timing for a lot of new things for free without any added mess to the codebase.

TL;DR: Added Profiler utility, tidier logs, no more SOUT and easier time measurement.