Added simple vulkan memory allocator/manager
Added intermediate memory allocator layer so that multiple buffers may be fit into a single allocation. This is due to the fact that as per Vulkan specifications, the logical structure for the data should not represent the physical structure in memory. That is, multiple buffers should be stored in same allocations. There is lot smart stuff that could be done with the memory allocator, but for now I implemented simple slice-based allocator which naively throws everything into large-ish dynamically allocated chunks.

With current settings, this reduced the overall allocation count from aroud 300-400 to 16 on len=200, SBS=2 cave and allowed me to generate a len=16k, SBS=2 cave without any errors. Generation time is still a constraint as the latter took almost four minutes to generate, but at least memory usage is much better under control now.